+++ test/UnitTestSocket/ConverterTest.cs
@@ -2,9 +2,9 @@
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 // Website: https://github.com/LongbowExtensions/
 
-namespace UnitTestTcpSocket;
+namespace UnitTestSocket;
 
-public class TcpSocketPropertyConverterTest
+public class ConverterTest
 {
     [Fact]
     public void UInt16Converter_Ok()
