+++ test/UnitTestSocket/UnitTestSocket.csproj
@@ -7,6 +7,8 @@
   </PropertyGroup>
 
   <ItemGroup>
+    <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="9.0.8" />
+    <PackageReference Include="Microsoft.Extensions.Options.ConfigurationExtensions" Version="9.0.8" />
     <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.*" />
     <PackageReference Include="xunit" Version="2.*" />
     <PackageReference Include="xunit.runner.visualstudio" Version="3.*">
@@ -21,17 +23,15 @@
 
   <ItemGroup>
     <ProjectReference Include="..\..\src\Longbow.Socket\Longbow.Socket.csproj" />
-    <ProjectReference Include="..\..\src\Longbow.TcpSocket\Longbow.TcpSocket.csproj" />
   </ItemGroup>
 
   <ItemGroup>
     <Using Include="Longbow.Socket.DataAdapters" />
     <Using Include="Longbow.Socket.DataHandlers" />
     <Using Include="Longbow.Socket.DataConverters" />
-    <Using Include="Longbow.TcpSocket" />
-    <Using Include="System.Diagnostics.CodeAnalysis"/>
-    <Using Include="Xunit"/>
-    <Using Include="Xunit.Abstractions"/>
+    <Using Include="System.Diagnostics.CodeAnalysis" />
+    <Using Include="Xunit" />
+    <Using Include="Xunit.Abstractions" />
   </ItemGroup>
 
 </Project>
